refactor: Extract clustering tests into sub-module
Move clustering integration tests from create_table.rs into a dedicated
create_table/clustering.rs sub-module, mirroring the column_mapping.rs
extraction. This keeps the main test file focused on core CreateTable
functionality while organizing feature-specific tests in their own files.

Testing
=======
Ran all the existing tests since this is a lift and shift
